Move LingBot Video to general reasoning
