+++ src/openworldlib/reasoning/general_reasoning/lingbot_video/prompt_templates.py
