--- src/openworldlib/reasoning/visual_reasoning/__init__.py
@@ -1,1 +1,0 @@
-"""Visual reasoning modules."""
