+++ test/test_lingbot_video_i2v.py
@@ -7,7 +7,7 @@ from diffusers.utils import export_to_video
 from PIL import Image
 
 from openworldlib.pipelines.lingbot_video.pipeline_lingbot_video import LingBotVideoPipeline
-from openworldlib.reasoning.visual_reasoning.lingbot_video import LingBotVideoReasoning
+from openworldlib.reasoning.general_reasoning.lingbot_video import LingBotVideoReasoning
 
 
 model_path = os.environ.get("LINGBOT_VIDEO_MODEL_PATH", "Robbyant/lingbot-video-dense-1.3b")
